chore: upgrade all deps and move to docz
