+++ docs/index.mdx
@@ -1,0 +1,14 @@
+---
+name: Home
+route: /
+order: 1
+---
+
+# Downshift
+
+This might be the future of the docs. But for now we're just using this for
+Cypress tests.
+
+For docs see the repo: [github.com/paypal/downshift](https://github.com/paypal/downshift)
+
+For examples see: [github.com/kentcdodds/downshift-examples](https://github.com/kentcdodds/downshift-examples)
