+++ docs/tests/combobox.mdx
@@ -1,0 +1,11 @@
+---
+name: Combobox
+route: /tests/combobox
+menu: Tests
+---
+
+import Combobox from './combobox'
+
+# Combobox
+
+<Combobox />
