+++ docs/tests/index.mdx
@@ -1,0 +1,7 @@
+---
+name: Tests
+route: /tests
+menu: Tests
+---
+
+This is where we render components for the Cypress tests.
