+++ package.json
@@ -20,17 +20,16 @@
     "test:flow": "flow",
     "test:flow:coverage": "flow-coverage-report",
     "test:build": "jest --projects other/misc-tests other/react-native",
-    "test:cypress:dev": "npm-run-all --parallel --race storybook cy:open",
-    "pretest:cypress": "npm run storybook:build --silent",
-    "test:cypress": "start-server-and-test storybook:serve http://localhost:6006 cy:run",
+    "test:cypress:dev": "npm-run-all --parallel --race docs:dev cy:open",
+    "pretest:cypress": "npm run docs:build --silent",
+    "test:cypress": "start-server-and-test docs:serve http://localhost:6006 cy:run",
     "cy:run": "cypress run",
     "cy:open": "cypress open",
     "build-and-test": "npm run build -s && npm run test:build -s",
-    "storybook": "start-storybook -p 6006 -c stories",
-    "storybook:build": "build-storybook -c stories",
-    "storybook:install": "cd stories && npm install",
-    "storybook:serve": "serve ./storybook-static -p 6006",
-    "setup": "npm install && npm run storybook:install && npm run validate",
+    "docs:build": "docz build",
+    "docs:dev": "docz dev",
+    "docs:serve": "serve ./.docz/dist --listen 6006 --single",
+    "setup": "npm install && npm run validate",
     "validate": "kcd-scripts validate lint,build-and-test,test:cover,test:ts,test:flow:coverage,test:ssr,test:cypress",
     "precommit": "kcd-scripts precommit"
   },
@@ -61,32 +60,33 @@
     "react": ">=0.14.9"
   },
   "dependencies": {
-    "compute-scroll-into-view": "^1.0.2",
+    "compute-scroll-into-view": "^1.0.9",
     "prop-types": "^15.6.0"
   },
   "devDependencies": {
-    "@storybook/react": "^3.4.7",
-    "@types/react": "^16.0.40",
-    "babel-plugin-macros": "^2.4.0",
+    "@types/react": "^16.4.14",
+    "babel-plugin-macros": "^2.4.2",
     "babel-preset-react-native": "^5.0.0",
     "cross-env": "^5.1.4",
-    "cypress": "3.0.1",
+    "cypress": "^3.1.0",
+    "cypress-testing-library": "^2.2.2",
+    "docz": "^0.12.5",
     "eslint-plugin-cypress": "^2.0.1",
-    "flow-bin": "^0.79.0",
-    "flow-coverage-report": "^0.5.0",
-    "jest-dom": "^1.3.1",
-    "kcd-scripts": "^0.43.0",
+    "flow-bin": "^0.82.0",
+    "flow-coverage-report": "^0.6.0",
+    "jest-dom": "^2.0.2",
+    "kcd-scripts": "^0.44.0",
     "npm-run-all": "^4.1.2",
     "preact": "^8.2.6",
     "preact-render-to-string": "^3.7.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
-    "react-native": "^0.56.0",
-    "react-test-renderer": "^16.3.2",
-    "react-testing-library": "^5.0.0",
-    "serve": "^10.0.0",
-    "start-server-and-test": "^1.4.1",
-    "typescript": "^2.7.2"
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "react-native": "^0.57.2",
+    "react-test-renderer": "^16.5.2",
+    "react-testing-library": "^5.2.0",
+    "serve": "^10.0.2",
+    "start-server-and-test": "^1.7.4",
+    "typescript": "^3.1.1"
   },
   "eslintConfig": {
     "extends": "./node_modules/kcd-scripts/eslint.js",
@@ -107,7 +107,7 @@
     "node_modules",
     "coverage",
     "dist",
-    "storybook-static",
+    ".docz",
     "typings"
   ],
   "repository": {
