--- stories/.babelrc
@@ -1,14 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "browsers": "ie >= 10"
-        }
-      }
-    ],
-    "@babel/preset-react"
-  ],
-  "plugins": ["babel-plugin-macros", "@babel/plugin-proposal-class-properties"]
-}
