--- stories/.eslintrc
@@ -1,18 +1,0 @@
-{
-  "extends": [
-    "../node_modules/kcd-scripts/eslint.js",
-  ],
-  "root": true,
-  "rules": {
-    "max-len": "off",
-    "no-console": "off",
-    "no-alert": "off",
-    "react/prop-types": "off",
-    "react/jsx-closing-bracket-location": "off",
-    "react/jsx-indent": "off",
-    "jsx-a11y/no-noninteractive-element-interactions": "off",
-    "prefer-const": "off",
-    "eqeqeq": "off",
-    "complexity": "off"
-  }
-}
