--- stories/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "stories",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "Kent C. Dodds <kent@doddsfamily.us> (http://kentcdodds.com/)",
-  "license": "MIT",
-  "dependencies": {
-    "apollo-boost": "^0.1.10",
-    "axios": "^0.18.0",
-    "glamor": "^2.20.40",
-    "glamorous": "^4.13.1",
-    "graphql": "^0.13.2",
-    "graphql-tag": "^2.9.2",
-    "match-sorter": "^2.2.3",
-    "react": "^16.4.1",
-    "react-apollo": "^2.1.6",
-    "react-dom": "^16.4.1",
-    "react-instantsearch": "^5.1.0",
-    "react-popper": "^0.7.2",
-    "react-tiny-virtual-list": "^2.1.6",
-    "react-virtualized": "^9.20.0"
-  },
-  "devDependencies": {
-    "babel-macros": "^2.0.0"
-  }
-}
